requirement.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-appconf==1.0.4
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-imagekit==4.0.2
djangorestframework==3.12.4
idna==3.2
pilkit==2.0
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
schedule==1.1.0
sendotp==0.0.4
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
